Optimizacion y refactorizacion de la gestion de facturas y pagos

+++ database/migrations/2024_08_04_094154_invoices_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('invoice_number', 10)->unique();
             $table->date('issue_date');
             $table->date('due_date');
-            $table->integer('total_amount');
-            $table->integer('pending_amount')->default(0);
-            $table->integer('total_paid')->default(0);
+            $table->decimal('total_amount', 10, 2);
+            $table->decimal('pending_amount', 10, 2)->default(0);
+            $table->decimal('total_paid', 10, 2)->default(0);
             $table->foreignId('client_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('created_by')->nullable()->constrained('users')->onDelete('set null');
